# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxe Store API – elegant adult products storefront
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * e.g. New, Bestseller, Limited
     * @nullable
     */
  badge?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  category: string;
  imageUrl?: string;
  badge?: string;
  inStock?: boolean;
  isFeatured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  badge?: string;
  inStock?: boolean;
  isFeatured?: boolean;
}

export interface StoreSummary {
  totalProducts: number;
  totalCategories: number;
  newArrivalsCount: number;
  featuredCount: number;
}

export interface CategorySummary {
  slug: string;
  label: string;
  productCount: number;
  /** @nullable */
  description?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};