# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxe Store API – elegant adult products storefront
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * e.g. New, Bestseller, Limited
     * @nullable
     */
  badge?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  createdAt: string;
}