# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxe Store API – elegant adult products storefront
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  badge?: string;
  inStock?: boolean;
  isFeatured?: boolean;
}